+++ src/js/modules/memoTrainingStats.js
@@ -9,10 +9,9 @@ import {
     take,
     select,
 } from 'redux-saga/effects';
-const config = require('../config');
 const memoTrainingUtils = require('../memoTrainingUtils');
+const letterPairTableUtils = require('../letterPairTableUtils');
 
-const rp = require('request-promise');
 const moment = require('moment-timezone');
 
 const FETCH_STATS = 'FETCH_STATS';
@@ -46,33 +45,6 @@ const initialState = {
     isOpenBo5Tooltip: false,
     isOpenAo5Tooltip: false,
     isOpenScoresComponentTooltip: false,
-};
-
-// FIXME module/memoTrainingResult.jsと重複している
-const fetchLetterPair = (userName) => {
-    if (userName === '') {
-        const ans = {
-            success: {
-
-                code: 200,
-                result: [],
-            },
-        };
-
-        return Promise.resolve(ans);
-    }
-
-    const options = {
-        url: `${config.apiRoot}/letterPair?userName=${userName}`,
-        method: 'GET',
-        headers: {
-            'Content-Type': 'application/json',
-        },
-        json: true,
-        form: {},
-    };
-
-    return rp(options);
 };
 
 function * handleFetchStats () {
@@ -127,21 +99,9 @@ function * handleFetchStats () {
         // 2文字→レターペアの配列
         // 種目選択するたびにロードするのは無駄だが、そんなに時間はかからないので問題ない見込み
         // FIXME ここも、module/memoTrainingResultと重複
-        const letterPairDict = {};
+        let letterPairDict = {};
         if (event === memoTrainingUtils.MemoEvent.mbld) {
-            const fetchedLetterPair = yield call(fetchLetterPair, userName);
-            const letterPairs = fetchedLetterPair.success.result;
-            for (let k = 0; k < letterPairs.length; k++) {
-                const letterPair = letterPairs[k];
-                const letters = letterPair.letters;
-                const word = letterPair.word;
-
-                if (letters in letterPairDict) {
-                    letterPairDict[letters].push(word);
-                } else {
-                    letterPairDict[letters] = [ word, ];
-                }
-            }
+            letterPairDict = yield call(letterPairTableUtils.fetchLetterPair, userName);
         }
 
         const payload = {
